Change name csv
- Change the generated name for the csv.

- Download only when you search.
